requirements.txt

Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9

